requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
googletrans==2.4.0
idna==2.8
imageio==2.6.1
imutils==0.5.3
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
nltk==3.4.5
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
pyparsing==2.4.6
pytesseract==0.3.1
python-dateutil==2.8.1
PyWavelets==1.1.1
regex==2020.1.8
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
urllib3==1.25.7
Wand==0.5.8
